fixed issues from partner review
