--- Corona-Warn-App/src/main/java/de/rki/coronawarnapp/submission/SymptomIndication.kt
@@ -1,7 +1,0 @@
-package de.rki.coronawarnapp.submission
-
-enum class SymptomIndication {
-    POSITIVE,
-    NEGATIVE,
-    NO_INFORMATION
-}
